+++ tcs-persistence/src/main/java/com/transformuk/hee/tis/tcs/service/model/PlacementDetails.java
@@ -45,7 +45,7 @@ public class PlacementDetails {
   private LocalDate dateTo;
 
   @Column(name = "placementWholeTimeEquivalent")
-  private BigDecimal wholeTimeEquivalent;
+  private Float wholeTimeEquivalent;
 
   private Long siteId;
